refactor: extract windmill-types crate from windmill-common
