+++ backend/windmill-common/src/workspace_dependencies.rs
@@ -386,7 +386,7 @@ impl WorkspaceDependenciesPrefetched {
 
         Box::pin(async {
             let r = if let Some(wdar) =
-                language.extract_workspace_dependencies_annotated_refs(code, runnable_path)
+                crate::scripts::extract_workspace_dependencies_annotated_refs(&language, code, runnable_path)
             {
                 tracing::debug!(workspace_id, ?language, "found explicit annotations");
