+++ backend/windmill-dep-map/src/lib.rs
@@ -88,7 +88,7 @@ pub fn extract_referenced_paths(
 ) -> Option<Vec<String>> {
     let mut referenced_paths = vec![];
     if let Some(wk_deps_refs) = language
-        .and_then(|l| l.extract_workspace_dependencies_annotated_refs(raw_code, script_path))
+        .and_then(|l| windmill_common::scripts::extract_workspace_dependencies_annotated_refs(&l, raw_code, script_path))
         .map(|r| r.external)
     {
         let l = language.expect("should be some");
